delphi: First try to fix callback dispatch stopping on disconnect
Callbacks could still get dispatched while the socket was already closed.
This resulted in a guaranteed timeout exception for getters called from
callback functions.

Try to use an additional mutex to serialize disconnection and callback
dispatch. But the problem is, this could deadlock due to an mutex ordering
problem with the socket mutex.