csharp: Avoid race condition between disconnect probe thread and disconnect function
There is a chance that the disconnect function locks the socket lock and
tires to join the disconnect probe thread while the disconnect probe thread
is about to enter the socket lock. This results in a dead lock.

Avoid this problem by not holding the socket lock in the disconnect probe
thread. This locking was not necessary as the disconnect probe thread only
exists while the socket is valid.

Also add a send lock to the socket for serializing send calls, which was
realized by the socket lock before.