java: Refactor bindings to include changes for MATLAB more easily
Split IPConnection into a Java specific IPConnection class and a common
IPConnectionBase class. The Java specific class implements callback
handling and signatures in the known way. The MATLAB specific class will
include changes required for the Listner handling of the MATLAB Java
interface. This includes that listener parameters are passed as an
instance of java.util.EventObject.