Avoid possible connection state race condition on connect
Most bindings use the IP Connection socket member to indicate the
connection state. This results in a small time window between the the
creation of the socket and a successful connect call. If the connect
call is going to fail then the IPConnection is going to report being
connected even if it isn't and will not be because connect() is going
to fail.

Avoid this time window by using a temporary variable. This also results
in the IP Connection not lying anymore about being connected if this is
not the case yet.