Skip .net related zip generation if mono compiler too new
Currently we can't easily build the C# .dll with the current
configurations if a new Mono compiler is used.

For now we just don't build the .zip if a proper mono version
is not available, otherwise the whole "generate_all" script
doesn't work.

In the future we will probably have to drop the support for older
.net versions.