c: Fix misaligned array accesses in callback wrappers.
GCC 9.1 upwards warns (correctly) if a pointer to a member of a packed struct is accessed.
Such code is only generated for arrays of types > 1 byte in callback wrappers.
Fortunately all those arrays are accessed anyway by the little endian conversion.
Assigning the converted values into an aligned array fixes the issue.