mqtt: Connect to broker first. Support pre/post-connect in init-file.
Connecting to the MQTT broker first allows sending the connect callback
for the first successful connection to the IP connection. To support
this, the init-file can now be in the format
{
 "pre-connect": {...}
 "post-connect": {...}
}
to be able to register the connected callback. Older init-files are
executed as post-connect, i.e. after the IP connection is established.