go: Insert '_' in constants between digit words.
Because constant names are in camel case in Go, constants like
'1 1 ms' were created as 11ms but have the meaning 1,1ms. This is
confusing and could collide with a constant '11 ms'. FlavoredName
now has camel_constant_safe, which will insert '_' between words
if the last word ends with a digit and the next starts with one.

The old constants are still generated and marked as deprecated if
they were released in an older version.