openhab: Add custom heartbeat mechanism.
The openHAB bindings now periodically call getIdentity of any
available device. If all devices attached to one bridge (i.e.
Brick Daemon) don't respond, the bindings assume that the connection
to the Brick Daemon was lost and try to reconnect. A custom
reconnect mechanism with exponential backoff (limited to at least
one attempt per 10 minutes) is used. If another request to a device
fails because of a timeout, the heartbeat mechanism is triggered
immediately.

This fixes unresponsive stacks connected over Ethernet/WiFi.