openhab: Rework outdoor weather station/sensor.
Using the sensor/station ID as (openHAB) UID and basing the
discovery mechanism on this ID results in the user having to
reconfigure channels, rules etc. every time the sensor/station
restarts.

Now the ID is a configuration parameter of the sensor/station,
that can be changed without creating a new (openHAB) device.
The discovery mechanism is removed, sensors/stations must be added
manually. This also fixes cache invalidation problems in the inbox.