Add device-identifier check for all API bindings, except Go and Rust
On the first setter/getter call of a device the bindings will call
get-identity internally to validate that the provided UID actually
matches the chosen device type. The result of this check is cached
and used on all subsequent setter/getter calls.

If a mismatch is detected the setter/getter call fails with the new
WRONG_DEVICE_TYPE error.

The same logic is applied to callbacks. Callbacks will only be handled
if the device-identifier matches the device-type. Otherwise callbacks
will be dropped silently.

Go and Rust support will be added in a later commit.