Switch all imports to full-absolute imports
Create generators module on the fly to allow for keeping the improper
directory structure while using a proper module structure to fix various
import problems due to sys.path hacking and incomplete-absolute imports.

This allows for the different generators to properly import each other.

This also requires to raise the required Python version to 3.5 in order
to use the importlib.util.module_from_spec function.