uc: Use custom printf implementation for logging and examples.
This removes the dependency to vsnprintf or it's custom implementation
that depended on itoa and utoa.

All log functions are modified to not automatically print a \n
as last character.

The printf implementation does not support zero-padding, grouping,
l-modifier or similar and no floats. Newlines (\n) are translated
to the platform specific newline character(s).