uc: Remove TF_CONST_STRING. Rename errors.inc.
The attempted RAM size optimization for the Arduino did not work,
as it requires additional code to read back the strings from flash.
This would complicate the printing code a lot, so it's better to
remove it entirely.

Disabling the tf_hal_strerror implementation should be sufficient.

Also rename errors.inc to error_cases.h as the Arduino IDE fails to
handle sources in non .h/.cpp (or similar) files.