javascript: Fix 64 bit integer support.
64 bit integers are only supported on newer nodejs versions
(>= v12.0.0 or v10.20.0). Check if the required functions
are available and return a (new) error code if not. This is
a breaking change in theory, however the old behaviour was to
just interpret the bytes as a double which is obviously broken.

This also adds the errorCallback parameter to the .on
method used for callback registrations to be able to communicate
the missing int64 support in case of a callback registration.