common: Add temporary hack for the EVSE bricklet.
common.py expects unreleased devices to always have an API version
of 2.0.0. While the EVSE Bricklet is unreleased, it is already in
the field as part of the WARP charger, so API changes should already
be versioned. Disable the check for now. This can be reverted when
the EVSE bricklet is released.