uc: Set UID in tf_unknown_create.
Not setting the UID here results in broadcasting all communication
to this port. This is the wanted behavior in hal_common_prepare, as
we want to enumerate all devices connected to a port.

However the unknown bricklet is also used to generically interact with
any CoMCU bricklet (i.e. to only implement the firmware flashing logic
once). In this case we only want to communicate with one Bricklet of
unspecified type, to for example not accidently attempt to flash an
isolator bricklet.

hal_common_prepare already passes the correct UID (0 or "1") for a
broadcast.