uc/streaming: Fix stream_out stream length confusion. Rename to _read.
The bug was a mix up of stream_length (the complete length of
the data to be streamed) and *ret_stream_length (the number of
items that were already transmitted) when checking whether the
stream is out of sync.

To be less confusing, rename ret_stream_length to ret_stream_read.