thermal-imaging: Fix high contrast ROI value ranges
The high contrast ROI constrains are different from the spot meter ROI
constrains. The data sheet is inconsistent. It seems that the spot meter
ROI minimum is 2x2 pixel and the high contrast ROI minimum is 1x1 pixel.
Due to typos in the data sheet the firmware currently implements a high
contrast ROI minimum size of 1x2 pixels. We decided to not change the
firmware, but adapt the documentation to the firmware.