rust: Add IPConnection::get/set_connection_timeout.
The connection timeout was hardcoded to 30 seconds, now it can be
changed before calling IPConnection::connect. The timeout is now
divided by the number of addresses a hostname resolved to, to make
sure the connection_timeout is not exceeded.

If the connection timeout is set to Duration::ZERO, the system's
default timeout is used for each connection attempt.